Use JIRA Rest API
