+++ .gitmodules
@@ -2,3 +2,7 @@
 	path = Modules/C--Toggl-Api-Client
 	url = https://github.com/johnbabb/C--Toggl-Api-Client.git
 	branch = master
+[submodule "Modules/JiraRestClient"]
+	path = Modules/JiraRestClient
+	url = git@github.com:dstj/JiraRestClient.git
+	branch = master
