+++ JiraTogglSync.Services/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Atlassian.SDK" version="2.5.0" targetFramework="net45" />
+  <package id="RestSharp" version="104.1" targetFramework="net45" />
 </packages>
